support BigInt schema type (64-bit int)
Implemented this change so that the page table can have offsets greater
than maxInt32.

All SQL integer values are stored as int64 in application code. This
allows us to handle both 32-bit and 64-bit values without adding
complicated logic. We are explicit about the type, as opposed to
simply using the int type, in order to support 64-bit integers on
32-bit systems.

When we serialize 32-bit integers before writing out to disk, we
ensure that the int64 value is within 32-bit range, and the value
is written out as a 32-bit integer, which saves space.

+++ cmd/csvimport/main_test.go
@@ -49,8 +49,8 @@ func TestGetDataTypes(t *testing.T) {
 				return nil, nil, errors.New("expected fetch for `sys_schema`")
 			}
 			return []*storage.Row{
-					{Vals: []interface{}{"author", "name", int32(storage.TypeVarchar)}},
-					{Vals: []interface{}{"author", "age", int32(storage.TypeInt)}},
+					{Vals: []interface{}{"author", "name", int64(storage.TypeVarchar)}},
+					{Vals: []interface{}{"author", "age", int64(storage.TypeInt)}},
 				},
 				[]*storage.Field{
 					{Column: "table_name"},
@@ -90,9 +90,9 @@ func TestCSVImport(t *testing.T) {
 				return nil, nil, errors.New("expected fetch for `sys_schema`")
 			}
 			return []*storage.Row{
-					{Vals: []interface{}{"author", "name", int32(storage.TypeVarchar)}},
-					{Vals: []interface{}{"author", "age", int32(storage.TypeInt)}},
-					{Vals: []interface{}{"author", "active", int32(storage.TypeBoolean)}},
+					{Vals: []interface{}{"author", "name", int64(storage.TypeVarchar)}},
+					{Vals: []interface{}{"author", "age", int64(storage.TypeInt)}},
+					{Vals: []interface{}{"author", "active", int64(storage.TypeBoolean)}},
 				},
 				[]*storage.Field{
 					{Column: "table_name"},
@@ -164,10 +164,10 @@ func TestCSVImport(t *testing.T) {
 	}
 
 	expected := [][]interface{}{
-		{"Person1", int32(10), true},
-		{"Person2", int32(20), false},
+		{"Person1", int64(10), true},
+		{"Person2", int64(20), false},
 		{"Person3", nil, true},
-		{"Person5", int32(15), false},
+		{"Person5", int64(15), false},
 	}
 
 	if !reflect.DeepEqual(expected, importedRows) {
